consolidate use of package:analyzer APIs
