+++ pkgs/dart_services/lib/src/pub.dart
@@ -4,18 +4,10 @@
 
 import 'dart:io';
 
-import 'package:analyzer/dart/analysis/utilities.dart';
-import 'package:analyzer/dart/ast/ast.dart';
 import 'package:path/path.dart' as path;
 import 'package:yaml/yaml.dart';
 
 import 'project_templates.dart' as project;
-
-/// Extract all imports from [dartSource] source code.
-List<ImportDirective> getAllImportsFor(String dartSource) {
-  final unit = parseString(content: dartSource, throwIfDiagnostics: false).unit;
-  return unit.directives.whereType<ImportDirective>().toList();
-}
 
 /// Flutter packages which do not have version numbers in pubspec.lock.
 const _flutterPackages = [
@@ -65,12 +57,4 @@ Map<String, String> packageVersionsFromPubspecLock(String templatePath) {
   });
 
   return packageVersions;
-}
-
-extension ImportDirectiveExtension on ImportDirective {
-  bool get dartImport => Uri.parse(uri.stringValue!).scheme == 'dart';
-
-  bool get packageImport => Uri.parse(uri.stringValue!).scheme == 'package';
-
-  String get packageName => Uri.parse(uri.stringValue!).pathSegments.first;
 }
