+++ pkgs/dart_services/pubspec.yaml
@@ -15,8 +15,8 @@ dependencies:
   dartpad_shared:
     path: ../dartpad_shared
   encrypt: ^5.0.3
-  google_cloud_ai_generativelanguage_v1beta: ^0.4.0
-  google_cloud_rpc: ^0.4.0
+  google_cloud_ai_generativelanguage_v1beta: ^0.5.1
+  google_cloud_rpc: ^0.5.1
   http: ^1.6.0
   logging: ^1.3.0
   meta: ^1.17.0
