+++ pkgs/dart_services/test/presubmit/analyzer_test.dart
@@ -2,11 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:dart_services/src/pub.dart';
+import 'package:dart_services/src/analyzer.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('pub', () {
+  group('analyzer', () {
     group('getAllImportsFor', () {
       test('empty', () {
         expect(getAllImportsFor(''), isEmpty);
